Move `BuildPlan` under `pkg` module in anticipation for `forc update`
